+++ LimeMeta/Logics/UserLogic.cs
@@ -456,7 +456,14 @@ public sealed class UserLogic : BaseLogic<User>
 /// </summary>
 public class LoginResult
 {
+    public Guid? UserId { get; set; }
+
     public string? Name { get; set; }
+
     public string? Token { get; set; }
+
+    public string? Email { get; set; }
+
+    public string? AvatarUrl { get; set; }
 }